Some refactoring of the WhitehallImporter::CreateEdition spec
These tests seemed to have a common battle with a verbose, repetitive
call to the `call` method under test. I've abstracted that out to a
method so the tests can be simpler. Along the way I've fixed an issue
where a context was outside the describe block and fixed some of the
more verbose code.