Run yarn upgrade
This updates various and many dependencies in a way that's not at all
scary about what the consequences are. This is done to address to
address a Github security alert for the minimalist dependency. I'm
pretty sure that dependency isn't actually bundled in the code we ship
but still nice to clear the alert.