Stop tests relying on unpredictable database ordering
These tests expect document.editions to return editions in a particular
ordering. PostgreSQL does not respect the ordering the author expects of
these and will opt for which ordering it finds most efficient -
unfortunately in most cases this will be the ordering the records are
inserted which makes these tests fail on rare occurrences.