Set time lets as time objects
These were somewhat surprising as string values as you'd normally
anticipate variables named in this way to be a form of a time object.

I used modifiers such as midday to set these to a fixed sub-second
precision. This allows these times to be converted to rfc3339 strings
and back without becoming different values.