Add auth_bypass_id column to editions table
This column is intended to be populated by a uuid that is created for
each edition. The value of this will be shared with the Publishing API
and used in the process to bypass authentication to the draft stack for
a particular content item.

This is introduced to replace the current approach where the documents
content id is used. This approach is flawed for the following reasons:

- It allows an old token to view later drafts of a piece of content
- It isn't practical to change the id in the event of a leak

In order for this system to be backwards compatible these
auth_bypass_id's are initially set-up to use the same algorithm as is
used currently to produce identical ids. Once this is applied we can
switch these to be based on UUID's generated at runtime.