Use auth_bypass_id from edition for auth bypass tokens
Now that we have an auth_bypass_id in each edition we can remove the
code to generate these from a content id and instead pull it straight
from an edition.

Now that generating a token is much simpler the purpose of the
PreviewAuthBypass seems rather unclear. I've decided this code makes
more sense in the edition model since it is already interconnected with
the auth_bypass_id attribute there.

I've included the content_id in the token too since we can't
authoritatively determine the content the token is for with an
auth_bypass_id (they can be transitory). I spent a while debating about
how we could include user uid as that really helps us now the source of
a leak. However this gets a big bogged down in whether that information
constitutes being PII and whether it should be encrypted. I decided
against adding that for now and something that can be added later if
there is a user need defined. The combination of issue time and
content-id should at least mean it can be narrowed down pretty closely
to a request and thus potentially a user if needed.