Use SCSS import for loading external files
Currently we have to make additional requests for these CSS files.
Looking back, we used to import these files in directly and with a
change in libraries we switched these to be additional requests without
comment [1].

My assumption is that this was done to deal with a limitation and this
is the intended approach. Doing this saves having to deal with some
additional stylelint alerts.