Make alt text an optional field for images
The accessibility team have decided that alt text should no longer be a
required field for images. This is because they have decided that this
field is frequently badly used and no alt text is better than bad alt
text.

Because this changed the criteria for whether content is previewable or
not a number of tests needed to be changed to reflect communication with
Publishing API and Asset Manager where it wasn't previously needed.

This change mirrors a similar application in Whitehall: https://github.com/alphagov/whitehall/pull/5756