Fix app no longer linting ruby
This app lost the ability to lint Ruby following [1]. This re-instates
this ability by restoring the Jenkinsfile to a simpler place.

To make this app more conventional with other GOV.UK apps I've moved
lint to be the first task in the Rakefile.

A consequence of switching to the default rake task is that brakeman
will run twice in CI, while not ideal this doesn't seem a huge pain
point as the task typically completes in under 10s.

[1]: https://github.com/alphagov/govuk-jenkinslib/pull/74