Spruce up lint.rake
I don't think anyone cares whether rubocop is formatted in clang or not
so we may as well use the default. This also moves the factory bot
linting after the JS/CSS one since it is less obvious form of linting
and I believe it makes more sense to group the standard lintings
together.