Extend healthcheck to optionally check S3
This adds a new check to verify that the app is able to connect to
the S3 backing service we use for ActiveStorage. In order to avoid
any coupling with the contents of the bucket, the check simply makes
a connection to determine whether an item exists.

This check is primarily for the purpose of automating deployments,
where we make a request to the endpoint to verify the app is running
at a basic level. I chose to make it an optional check on the current
/healthcheck endpoint for a few reasons:

- Some user journeys are still possible even if the app cannot
connect to S3, so we should not say the app is unavailable.

- Amazon charges per-request to S3 [1], so we should avoid making lots
of them (currently, at least one per machine, every 10s).

- The connection to S3 is over HTTP and thus not persistent, which
means there is additional delay and likelihood of transient failure.

[1]: https://aws.amazon.com/s3/pricing/