Replace reorderable component with new gem version
This change removes the local reorderable component in favour of the global gem version.

This is done to bring the styling and functionality in line with the rest of the publishing components.

This had to be done in one commit as the original component was causing the gem component JS to fail.

There are some styling changes made:
- The up/down buttons is now secondary solid grey where it was secondary outlined before
- The reorderable list items now have a border around them
- The numbering has been removed.

https://trello.com/c/9CQm5A0k/251-update-publishing-tools-to-use-new-reorderable-component