Resolve ActiveStorage deprecations
The *_after_upload methods have been deprecated. I've adopted the
equivalent for creating now called create_and_upload.

For `build` there isn't an equivalent method as it is unsafe to upload
files to storage before persisting them to the database. To replicate
our past behaviour I've switched the build and upload to be a 2-step
process. The safety concerns of uploading before build aren't relevant
for a test environment.