Use inline_svg gem to import SVG files
This replaces the approach of using views for inline SVG files. Instead
we use the inline_svg gem which allows the SVG files to be placed in the
assets directory and then imported into views at run time. Doing this
allows us to abstract the HTML concerns (such as class and aria-hidden)
from the SVG and handle it in the view.

I've moved the files from the view directory to assets directory, I
adjusted the naming convention from snake_case to kebab-case to reflect
the existing convention in asset files.

The motivation for this change is the Rails deprecation warnings of:

DEPRECATION WARNING: Rendering actions with '.' in the name is deprecated: components/markdown_editor/_heading_two.svg (called from block in _app_views_components__markdown_editor_html_erb__1670259127403376712_135880 at /Users/kevindew/govuk/content-publisher/app/views/components/_markdown_editor.html.erb:97)