Fix remaining rubocop-govuk 4.0.0 violations
Most of these were fine to fix, the one that proved to be a bit of a
pain was "RSpec/StubbedMock" which forbids mixing mocking returns with
assertion that mocks are called. I think that fixing these produced
better code that is easier to read but it doesn't seem a total slam
dunk.