Resolve unconventional Requirements::Form::FileAttachmentMetadataChecker
This class previously had a different `.call` interface to the other
classes that implement Requirements::Checker, this was needed due to the
passing of a strong parameters object that hadn't had it's attributes
permitted (for some reason that I don't fully understand `.to_h` is
called on the params).

By taking the more conventional route of permitting attributes this
problem can be resolved and the unconventional code removed.