Migrate from Jasmine gem to jasmine-browser-runner
The Jasmine Ruby gems have been [deprecated][1] and are no longer supported. And whilst they still work with Rails 6 apps, they [aren't compatible with Rails 7][2] which is now becoming a blocker.

This commit implements the `jasmine-browser-runner` npm package, which is now the recommended approach for running Jasmine tests in the browser, and removes the deprecated Ruby gems.

This follows the example set by alphagov/govuk_publishing_components#2625

[1]: https://github.com/jasmine/jasmine-gem#discontinued
[2]: https://github.com/alphagov/govuk_publishing_components/issues/2621#issuecomment-1048007995