Add attachment arguments
In order to present a link to the new accessible format request form, attachments will need to provide the attachment_id and the owning_document_content_id.

Initially, the new flow of requesting accessible formats will be a pilot and only available for documents
owned by a few organisations. This code will not need to be changed when the change is
released to all orgs.

Co-authored-by: Peter Hartshorn <peter.hartshorn@digital.cabinet-office.gov.uk>