Update JS tests for no jQuery
This removes both direct references to jQuery and the use of
jasmine-jquery matchers.

There was one test I really struggled to make work, the Markdown
component focus test. I couldn't work out a good way of doing it
that wasn't contrived (document.activeElement doesn't seem to work) so I
eventually decided it wasn't a huge loss to drop it. The code is already
executed as part of other tests, just this side effect is no longer
actively covered.