Remove non-functional Notify rake task
Running this rake task produces an error:

```
Notifications::Client::BadRequestError: ValidationError: template_id is not a valid UUID
  lib/tasks/notify.rake:12:in `block (2 levels) in <main>'
    ActionMailer::Base.mail(**params).deliver_now
```

This seems to have occurred when this application switched to the
mail-notify gem in https://github.com/alphagov/content-publisher/pull/1926

I discovered this problem today when I tried to use the task.

A fix for it would be to add template_id as an argument as part of the
params, but given it's been over 2 years since this was broken it seems
more prudent to remove the code than fix it.