Enable aliases when opening document_types.yml
This YAML file makes use of the aliases feature and is raising
exceptions with bootsnap 1.16.0 when it is opened. eg:

```
     Failure/Error: hashes = YAML.load_file(Rails.root.join("config/document_types.yml"))["document_types"]

     Psych::BadAlias:
       Unknown alias: publication
```

I'm a bit baffled by exactly why this is caused by a bootsnap update. I
would expect that this error should have occurred when we updated to
Ruby 3.1 and thus started using Psych 4.

However, from some debugging I can see that bootsnap seems to overload
Psych which may explain why this has appeared unusually.