Remove delay from Redis reconnection attempts
The commit c9bb50d7a21b0b92181c2a974fe237a304053f0a changed the
`reconnect_attempts` parameter given to
ActiveSupport::Cache::RedisCacheStore from an integer to an array. This
was done to support a Sidekiq upgrade from v5 to v6.

While this looked like a safe and reasonable change to make, it seems to
have caused issues when it comes to upgrading the associated gems
sentry-sidekiq and sidekiq-scheduler.

Test runs were failing with error messages like this:

```
Failure/Error: delegate :clear, :exist?, :middleware, to: :cache

NoMethodError:
  undefined method `to_i' for [0.1, 0.1, 0.1]:Array

        options[:reconnect_attempts] = options[:reconnect_attempts].to_i
                                                                   ^^^^^
  Did you mean?  to_s
                 to_a
                 to_h
```

It looks like something, somewhere, was still expecting the
`reconnect_attempts` parameter to be an integer, rather than an array of
floats.

Therefore, I've switched this back to an integer and tests seem to be
passing again.

As I understand it, the impact of this change is that reconnection
attempts will no longer be delayed. Presumably they'll happen
immediately, one after the other, until the retry count has been
exhaused.

I'm not aware of any particular reason why the delays were required or
if they were needed in practice. So I'm going to tentatively put this
change through, and will keep an eye on Sentry issues to see if errors
arise as a consequence.