Ignore stylelint rule for special case
The loading spinner SCSS file is slightly unusual because it defines
lots of (almost identical) CSS animations which are required to make the
spinner animation work.

This isn't a 'normal' SCSS file, and it should not be used as an example
of how to write SCSS files in this project.

The `declaration-block-single-line-max-declarations` rule had already
been ignored. But Stylelint also started complaining about
`keyframe-block-no-duplicate-selectors` after bumping
stylelint-config-gds from 0.2.0 to 0.3.0. So I've ignored that too, just
for this particular file.