Move google tag manager snippet
In order to be consistent with the other publishing apps we have moved
the google tage snippet into a partial. This matches the way this was
done in all the other publishing apps owned by the publishing experience
team.

Google Tag Manager (GTM) can be included on all pages of the app by configuring the following environment variables:

GOOGLE_TAG_MANAGER_ID
GOOGLE_TAG_MANAGER_AUTH
GOOGLE_TAG_MANAGER_PREVIEW
All three variables are optional.

If none are set, GTM will not be included on the page.

If only GOOGLE_TAG_MANAGER_ID is set, the default environment of the GTM container will be used.

If GOOGLE_TAG_MANAGER_AUTH and GOOGLE_TAG_MANAGER_PREVIEW are set, the specified GTM container environment will be used. This is used for managing the rollout of changes to integration so they can be tested before rolling out to production.