Remove 'scrub access limited' SQL
The original intention was to keep this as the main copy of the SQL
sanitisation script (the script that is applied to data when it
is copied from Production to Integration in the nightly env sync).
A code comment deleted here refers to https://github.com/alphagov/env-sync-and-backup/blob/master/sanitising-sql/sanitise_content_publisher_production.sql,
which is the 'copy' of this main copy, and should be kept in sync.

The env sync process has since moved from env-sync-and-backup to
govuk-helm-charts. The SQL script is referred to here:
https://github.com/alphagov/govuk-helm-charts/blob/61d36eeb6339c13600f1c5cecc725a71ff593053/charts/db-backup/values.yaml#L329

...and the SQL script it is referring to lives here:
https://github.com/alphagov/govuk-helm-charts/blob/main/charts/db-backup/scripts/content-publisher.sql

There is currently a TODO comment in that repo:

```
-- TODO: Fix the CI in alphagov/content-publisher to pull in this script
-- (perhaps via git submodule) instead of running tests against its own copy.
```

So let's do that.

This commit removes our local copy of the SQL sanitisation, and
pulls in the remote copy instead (the version that is actually
used in the env sync job, which is far more useful).