Start building a `DocumentExport` class
We will call this class from a rake task, in a later commit.
The class will be responsible for fetching all relevant documents
in Content Publisher and constructing an output / payload for
each of them, ready for importing into Whitehall. This will
enable us to retire Content Publisher.

In this first commit, we've built the behaviour that fetches the
relevant documents. We've decided not to include documents that
are only in a draft state, as the complexity in migrating drafts
(and particularly 'live with draft' documents) is high relative
to the value of delivering the feature (there are no drafts in
Content Publisher newer than early 2024).

We've also decided to not include 'removed' documents (which would
map to Unpublished-with-redirect state in Whitehall) for the same
reason: complexity in reconstructing the document in Whitehall
for little gain (only 25 documents affected) where Publishing API
will continue to host the redirects even without us migrating
those documents. NB, all but 3 have an alternative URL (redirect);
the other 3 are just 'gone' (410):
- /government/news/pro-direct-pays-charity-after-9-years-of-waste-packaging-errors
- /government/news/new-cma-campaign-urges-firms-to-compete-not-cheat
- /government/news/content-to-be-confirmed

We've also decided to build this exporter in the Whitehall
Migration folder which has an 'import' equivalent. This is a
holding place and we may move it later if it makes sense to do so.