Include created_by/last_edited_by info in the hash
We want to preserve the "created by" and "last updated by" data for Whitehall,
but we're comfortable dropping the 'author' of each of the intermediary changes,
which will [default "User (removed)" - see https://github.com/alphagov/whitehall/blob/9bd6e9f42523a8c21ca430e0fcde14bc31b74ef0/app/components/admin/editions/audit_trail_entry_component.rb#L20)
in the Whitehall UI. The alternative would have beento dig into every MetadataRevision to pull out
the `created_by_id`, and map that to a user, and map that to the particular
changenote associated with the MetadataRevision, which we felt was a lot of work
for little benefit, given we're unlikely to need to refer to that information.

(We can always retrieve the full changenote info from a Content
Publisher backup if we absolutely have to).