Add 'document_history' property
Whitehall stores change notes and editorial remarks _per edition_,
yet we're only planning to migrate the latest (published) edition
out of Content Publisher, for simplicity. So we imagine this
property will be used to form one large 'editorial remark' in
Whitehall (and potentially as one large entry in the public
changenote history too).

The TimelineEntry class is doing the heavy lifting here. I've
generally followed the same logic as _content_publisher_entry.html.erb
in deciding which properties to expose.

The way we're testing this behaviour (by layering edition upon
edition in a unit test) doesn't appear to have been done before,
as we were getting an Uninitialized Constant error on
`RevisionUpdater::Image`. Adding a `require_relative` to the
`RevisionUpdater` class worked around that.