Add 'change_notes' property
The 'document_history' property wasn't exposing any public
changenotes. We'll revisit document_history in the next commit.

I did consider rolling out my own change_notes implementation:

```
  def self.change_notes(document)
    change_notes = document.editions.map do |edition|
      change_note = edition.revision.metadata_revision.change_note
      if change_note.present?
        { change_note:, created_at: edition.created_at, created_by: User.find(edition.created_by_id).email }
      end
    end
    change_notes.compact
  end
```

...but on closer inspection, there is a PublishingApiPayload
`History` class that does exactly what we want and is already
fully tested:
https://github.com/alphagov/content-publisher/blob/ef1ee481f5309e6fd010967231ff5b4723db60ca/spec/lib/publishing_api_payload/history_spec.rb#L1

It is enough for our tests to check that that class is called.

That class also handles the backdating of `first_published_at`,
so we're fixing that whilst we're here.