Rename `document_history` to `internal_history` & drop backdating
In the initial implementation of this, we'd wrongly assumed that
public changenotes were included in the document history displayed
to publishers in the UI. We've since defined a separate
`change_notes` method for capturing that. That new method also
includes the 'backdating' information, and we've updated
`first_published_at` to accommodate backdating too, so it's no
longer required to be surfaced here.

What is still needed is some means of exposing 'internal' change
history, i.e. internal notes / editorial remarks, but also the
option of carrying over update/publish dates of each revision and
the author of each.

So in this commit we've renamed the method to better describe the
new scope - `internal_history` and have dropped any complexity
around backdating.